Added Chris' checkCurrentLrgStatus function to the comfy_BED_web script, made tests
