+++ comfy_BED/comfy_BED.py
@@ -9,7 +9,7 @@ import logging
 import requests
 import json
 
-from comfy_BED_web import getLrgFromWeb
+from comfy_BED_web import checkCurrentLrgStatus, getLrgFromWeb
 
 
 # load arguments
@@ -72,7 +72,6 @@ def getArgs():
         Select genome build from 'GRCh37' or 'GRCh38'. Defaults to GRCh37.
         '''
     ))
-
     return parser.parse_args()
 
 
@@ -84,39 +83,6 @@ def setUpLogs(args, now):
     '''
     log_filename = now.strftime("%Y-%m-%d") + "_comfy_BED" + ".log"
     logging.basicConfig(filename=log_filename, level=logging.DEBUG)
-
-
-def checkCurrentLrgStatus(lrg_id):
-    '''
-    Checks the CURRENT status of the user-provided LRG ID on LRG website, returns information to the BED header and log file
-    'Public' LRGs have a 'fixed annotation' section which has been fully finalised
-    'Pending' LRGs do NOT have a finalised 'fixed annotation' section
-    WARNING: User-provided XMLs contain no information as to whether they were public or private at time of download
-    The end user should always download their XMLs very shortly before use
-    '''
-    #get data from webservice
-    url_p1 = "https://www.ebi.ac.uk/ebisearch/ws/rest/lrg/entry/"
-    url_p3 = "?fields=status&format=json"
-    url_full = url_p1 + str(lrg_id) + url_p3
-    logging.info("Checking status with webservice: " + str(url_full))
-    data_return = requests.get(url_full)
-    parsed_data_return = data_return.json()
-
-    # parse the returned data, return status and log message
-    lrg_status_return = parsed_data_return['entries'][0]['fields']['status'][0]
-    if lrg_status_return == "public":
-        lrg_status_message = "The LRG is currently marked 'public' on the LRG website: note that the user-provided file could have been downloaded before the LRG going public"
-    else:
-        if lrg_status_return == "pending":
-            lrg_status_message = "The LRG is currently marked 'pending' on the LRG website: the fixed annotation is not yet finalised, so it should be interpreted with caution"
-    assert (lrg_status_return == "public") or (lrg_status_return == "pending"), "The LRG status could not be resolved as public or private"
-    if (lrg_status_return != "public") and (lrg_status_return != "pending"):
-        logging.error("The LRG status could not be resolved as public or pending")
-        lrg_status_message = "ERROR: The LRG status could not be resolved with the webservice as public or pending"
-
-    logging.info("LRG status is: " + lrg_status_return)
-    logging.info(lrg_status_message)
-    return lrg_status_return, lrg_status_message
 
 
 def getLrgExons(transcript, lrg_id):
